Fix OAK-5

+++ src/oracle/IMarketBorrowRateFeed.sol
@@ -4,5 +4,5 @@ pragma solidity 0.8.24;
 /// @title IMarketBorrowRateFeed
 interface IMarketBorrowRateFeed {
     /// @notice Returns the market borrow rate
-    function getMarketBorrowRate() external view returns (uint256);
+    function getMarketBorrowRate() external view returns (uint128);
 }
